fix: distributed components to folder
