+++ src/Components/MainPage/Banner/Banner.module.scss
