+++ src/Components/ShopPage/Categories/Categories.module.scss
@@ -1,0 +1,58 @@
+@import '@/styles/mixins';
+@import '@/styles/variables';
+
+.wrapper {
+	display: flex;
+	flex-direction: column;
+	width: 30%;
+}
+
+.title {
+	@include interBold(18px, $second-color);
+}
+
+.buttonBlock {
+	display: flex;
+	flex-direction: column;
+	align-items: start;
+}
+
+.categories {
+	@include interBold(18px, $second-color);
+	margin-bottom: 27px;
+}
+
+.btn {
+	@include interRegular(14px, $second-color);
+	border: none;
+	cursor: pointer;
+}
+
+@media (max-width: 768px) {
+	.wrapper {
+    display: none;
+  }
+	.visible {
+		width: 50%;
+		padding: 20px;
+		visibility: visible;
+		position: absolute;
+		background: $main-color;
+		z-index: 3;
+		margin-top: 120px;
+		border-radius: 15px;
+		border: 1px solid $second-color;
+	}
+	.notVisible {
+		visibility: hidden;
+		width: 0;
+	}
+	.title {
+		@include interBold(16px, $second-color);
+	}
+}
+@media (max-width: 500px) {
+	.title {
+		@include interBold(12px, $second-color);
+	}
+}
