+++ src/UI/ArrowButton/ArrowButton.module.scss
@@ -1,0 +1,8 @@
+@import '@/styles/mixins';
+@import '@/styles/variables';
+
+.arrow {
+  border:none;
+  cursor: pointer;
+  padding: 5px;
+}
