+++ src/UI/BuyButton/Button.module.scss
@@ -1,0 +1,59 @@
+@import '@/styles/mixins';
+@import '@/styles/variables';
+
+.btn {
+  @include interBold(16px, $second-color);
+	display: flex;
+	align-items: center;
+	justify-content: space-between;
+	cursor: pointer;
+	color: $second-color;
+	background: $btn-color;
+	border: none;
+	border-radius: 15px;
+	padding: 20px;
+	.image {
+		margin-left: 20px;
+	}
+}
+.filter {
+  @include interBold(16px, $second-color);
+	display: none;
+	align-items: center;
+	justify-content: space-between;
+	cursor: pointer;
+	color: $second-color;
+	background: $btn-color;
+	border: none;
+	border-radius: 15px;
+	padding: 20px;
+	.image {
+		margin-left: 20px;
+	}
+}
+.sortBy {
+  @include interBold(16px, $second-color);
+	display: flex;
+	align-items: center;
+	justify-content: space-between;
+	cursor: pointer;
+	color: $main-color;
+	background: $btn-color;
+	border: none;
+	border-radius: 15px;
+	padding: 20px;
+	.image {
+		margin-left: 20px;
+	}
+}
+
+@media(max-width: 768px) {
+	.btn {
+		@include interBold(14px, $second-color);
+		padding: 11px;
+		display: flex;
+	}
+	.filter {
+		display: flex;
+	}
+}
