--- src/app/components/ShopPage/Categories/RangePrice/RangePrice.module.scss
@@ -1,30 +1,0 @@
-@import '../../../../styles/mixins';
-@import '../../../../styles/variables';
-.rangePrice {
-  display: flex;
-  padding: 20px 0;
-  gap: 20px 0;
-  flex-direction: column;
-}
-.slider{
-  width: 90%;
-  margin: 0 auto;
-}
-.rangePriceParagraph {
-  @include interBold(18px, $second-color);
-  margin: 0; 
-  padding: 0;
-}
-.rangeSpan {
-  @include interRegular(14px, $second-color)
-}
-@media (max-width: 768px) {
-  .rangePriceParagraph {
-    @include interBold(16px, $second-color);
-  }
-}
-@media (max-width: 500px) {
-  .rangePriceParagraph {
-    @include interBold(12px, $second-color);
-  }
-}
