--- src/app/components/ShopPage/Products/Card/ArrowButton/ArrowButton.module.scss
@@ -1,8 +1,0 @@
-@import '../../../../../styles/mixins';
-@import '../../../../../styles/variables';
-
-.arrow {
-  border:none;
-  cursor: pointer;
-  padding: 5px;
-}
