--- src/app/components/ShopPage/Products/Card/RatingStar/RatingStart.module.scss
@@ -1,33 +1,0 @@
-@import '../../../../../styles/mixins';
-@import '../../../../../styles/variables';
-
-.ratingWrapper {
-  display: flex;
-  width: 70%;
-  min-width: 100px;
-  border-radius: 8px;
-  margin: 0 auto;
-  align-items: center;
-  justify-content: space-between;
-}
-.ratingStars{
-  width: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: space-around;
-}
-.star{
-  margin: 7px 5px;
-}
-
-@media (max-width: 768px) {
-  .ratingWrapper {
-    width: 80%;
-  }
-}
-
-@media (max-width: 500px) {
-  .ratingWrapper {
-    width: 100%;
-  }
-}
