--- src/app/components/ShopPage/Products/SearchInput/SearchInput.module.scss
@@ -1,36 +1,0 @@
-@import '../../../../styles/mixins';
-@import '../../../../styles/variables';
-.searchWrapper {
-	position: relative;
-	width: 100%;
-	border: 1px solid $second-color;
-	border-radius: 15px;
-	margin-bottom: 20px;
-}
-.searchInput {
-	padding-left: 40px;
-	@include interRegular(14px, $second-color);
-	opacity: 0.6;
-	width: 100%;
-	outline: none;
-	height: 50px;
-	border: none;
-}
-.searchImage {
-	position: absolute;
-	top: 15px;
-	right: 40px;
-}
-@media (max-width: 768px) {
-	.searchWrapper {
-		width: 90%;
-	}
-	.searchInput {
-		height: 30px;
-		display: flex;
-	}
-	.searchImage {
-		top: 7px;
-		right: 40px;
-	}
-}
