--- src/app/pages/shop/page.module.scss
@@ -1,116 +1,0 @@
-@import '../../styles/mixins';
-@import '../../styles/variables';
-
-.wrapper {
-	width: 100%;
-	background: $main-color;
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-}
-.headerImage {
-	background: url('/image/solarpanel.jpg');
-	color: $main-color;
-	width: 1140px;
-	height: 122px;
-	display: flex;
-	align-items: center;
-	background-size: cover;
-	margin-bottom: 40px;
-	.title {
-		padding-left: 70px;
-	}
-}
-.mainWrapper {
-	display: flex;
-	justify-content: space-between;
-	width: 1140px;
-}
-
-.buttonBlock {
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	margin-bottom: 20px;
-}
-.sortBy {
-	display: flex;
-	align-items: center;
-	span {
-    margin-right: 10px;
-	}
-	.filterProduct {
-		padding: 10px 5px;
-		border-radius: 5px;
-	}
-}
-
-.products {
-	width: 70%;
-	display: flex;
-	flex-direction: column;
-}
-
-.cards {
-	display: flex;
-	justify-content: space-between;
-	width: 100%;
-	flex-wrap: wrap;
-	gap: 6px 0;
-	flex-wrap: wrap;
-}
-@media (max-width: 1140px) {
-	.mainWrapper {
-		width: 100%;
-		display: flex;
-		justify-content: space-between;
-	}
-}
-
-@media (max-width: 927px) {
-	.mainWrapper {
-		display: flex;
-		width: 100%;
-		min-width: 100%;
-		justify-content: space-between;
-	}
-	.products {
-		display: flex;
-		align-items: center;
-		justify-content: space-between;
-		width: 70%;
-		min-width: 70%;
-		margin: 0 auto;
-		position: relative;
-	}
-	.cards {
-		width: 100%;
-		display: flex;
-		align-items: center;
-		justify-content: space-around;
-	}
-	.close {
-		display: block;
-	}
-}
-@media (max-width: 768px) {
-	.products {
-		width: 100%;
-	}
-	.buttonBlock {
-		display: flex;
-		width: 100%;
-		align-items: center;
-		justify-content: space-around;
-		padding: 10px 0;
-	}
-	.sortBy {
-		flex-direction: column;
-		span {
-			display: none;
-		}
-	}
-}
-
-@media (max-width: 500px) {
-}
