+++ src/app/shop/[id]/page.module.scss
@@ -1,0 +1,9 @@
+@import '@/styles/mixins';
+@import '@/styles/variables';
+
+.wrapper {
+	width: 1140px;
+	margin: 0 auto;
+	display: flex;
+	background-color: $main-color;
+}
